requirements.txt
alabaster==0.7.13
allure-pytest==2.13.2
allure-python-commons==2.13.2
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
attrs==23.1.0
Babel==2.12.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
coverage==7.3.3
data==0.4
DateTime==5.2
decorator==5.1.1
docutils==0.18.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Faker==21.0.0
funcsigs==1.0.2
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
h11==0.14.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
latex==0.7.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
mypy==1.5.1
mypy-extensions==1.0.0
numpy==1.26.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
pluggy==1.3.0
Pygments==2.16.1
PySimpleGUI==4.60.5
PySocks==1.7.1
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.14.0
shutilwhich==1.1.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
tempdir==0.7.1
tex==1.8
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
types-requests==2.31.0.3
types-urllib3==1.26.25.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
wsproto==1.2.0
zipp==3.16.2
zope.interface==6.0
